Model_Backend/requirements.txt
# ── Web server ──────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ── Environment ─────────────────────────────────
python-dotenv>=1.0.0

# ── ML / inference ─────────────────────────────
ultralytics
opencv-python-headless
Pillow
numpy
torch
torchvision
transformers>=4.35.0

# ── PDF manifest extraction ────────────────────
pdfplumber>=0.11.0
fpdf2

# ── Data / utilities ───────────────────────────
matplotlib
pandas>=2.0.0
scikit-image>=0.22.0
# -- VLM invoice extraction & async HTTP ----------
pymupdf>=1.24.0
httpx>=0.27.0

# ── External services (Supabase, Pinata, VLM) ──
requests>=2.31.0

# ── Encryption (audit log security) ────────────
cryptography>=42.0.0
